Reserve enough space for lvm metadata when computing PV size usable for LVs (#500431)
This is port of patch which went in RHEL 4.9. Up to now we reserved one more
PE for metadata if PV size was equal to clamped size (too narrow). With the
patch, we reserve one more PE if clamping itself would reserve less than 1MB.

+++ iw/lvm_dialog_gui.py
@@ -827,11 +827,9 @@ class VolumeGroupEditor:
                 pvsize = pvreq.getActualSize(self.partitions, self.diskset)
                 # have to clamp pvsize to multiple of PE
                 clampedSize = lvm.clampPVSize(pvsize, curpe)
-                if pvsize == clampedSize:
-                    # This is a corner case were we say that the available size of the VG
-                    # is the total size of the paritition.  Given that the pvs need metadata
-                    # it is better to be on the safe side and give ourselves on pe of extra
-                    # space.
+                if long(pvsize) == clampedSize:
+                    # If clamping reserves only less than 1MB for lvm metadata,
+                    # reserve one more PE.
                     clampedSize = clampedSize - (curpe / 1024)
 
                 availSpaceMB = availSpaceMB + clampedSize
